backend/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.3
python-dotenv==0.21.0
eventlet==0.33.3
yt-dlp==2023.11.16
requests==2.31.0
python-telegram-bot==20.3
pyperclip==1.8.2
pywin32==306; sys_platform == 'win32'
user-agents==2.2.0
Jinja2==3.1.2
python-engineio==4.5.1
python-socketio==5.8.0
SQLAlchemy==2.0.23
pymysql==1.1.0
PyJWT==2.8.0
cryptography==41.0.3
python-dateutil==2.8.2
pytz==2023.3
celery==5.3.4
redis==4.5.5
sphinx==7.0.1
sphinx-rtd-theme==1.3.0